UI/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
attrs==25.1.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.3.0
pyserial==3.5
pyserial-asyncio==0.6
pywin32==308
websockets==15.0
yarl==1.18.3
